chat_app/app/requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-language==2.13.4
google-cloud-translate==3.15.4
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.7
importlib_metadata==8.2.0
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
proto-plus==1.24.0
protobuf==5.27.2
ptyprocess==0.7.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
rapidfuzz==3.9.4
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
streamlit==1.37.0
streamlit_arabic_support_wrapper==1.1
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.0
toolz==0.12.1
tornado==6.4.1
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
watchdog==4.0.1
xattr==1.1.0
zipp==3.19.2
